circle-telegram-bot/requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
aiohttp==3.9.5
aiofiles==24.1.0
asyncio-throttle==1.0.2

# Для логирования
loguru==0.7.2

# Для работы с датами
python-dateutil==2.9.0

# Для HTTP запросов к backend API
httpx==0.27.0

# Для работы с изображениями (если понадобится)
Pillow==10.4.0 